requirements.txt
PyYAML
zmq
IPython
pyelftools
scapy
pycparser
keystone-engine
nose
